refactor: перенос стилей и типов
