+++ src/app/Theme.ts
@@ -1,0 +1,8 @@
+export const theme = {
+  primary: '#4C72B0',
+  secondary: '#DD8452',
+  background: '#f9f9f9',
+  text: '#333',
+  grid: '#e0e0e0',
+  tooltipBg: 'rgba(255, 255, 255, 0.9)',
+};
